Added unique puzzle checker
The verify function itself solves the puzzle with backtracking, so I've replaced the backtrack function with verify in the puzzlegen. If verify returns false, puzzlegen recursively calls itself to try again on the same grid. It'll stop calling recursively when a unique puzzle is generated, at which point a message is sent confirming this to standard output.